refactor(color-clash): split large component into smaller files
- Split logic from color-clash-game.component.ts into separate model and animation files
- Created color-clash-button.model.ts and color-clash-random-items.model.ts for better organization
- Moved animation logic to color-clash-item.animation.ts
- Renamed color-clash-interface.ts for consistency
